requirements.txt
anthropic
python-docx
watchdog
requests
python-dotenv
